Added UnitTests for projectLoader
